requirements.txt
opencv-python
ultralytics
face-recognition
numpy
Pillow
deep-sort-realtime